+++ R/kable_styling.R
@@ -79,8 +79,10 @@ kable_styling <- function(kable_input,
   kable_format <- attr(kable_input, "format")
 
   if (!kable_format %in% c("html", "latex")) {
-      message("Currently generic markdown table using pandoc is not supported.")
-      return(kable_input)
+    warning("Please specify format in kable. kableExtra can customize either ",
+            "HTML or LaTeX outputs. See https://haozhu233.github.io/kableExtra/ ",
+            "for details.")
+    return(kable_input)
   }
   if (kable_format == "html") {
     if (is.null(full_width)) {
@@ -169,6 +171,7 @@ htmlTable_styling <- function(kable_input,
 
   out <- as_kable_xml(kable_xml)
   attributes(out) <- kable_attrs
+  if (!"kableExtra" %in% class(out)) class(out) <- c("kableExtra", class(out))
   return(out)
 }